Delete Authentication Section
I think it's best for us to punt on this section right now. There are too many conversations up in the air regardings SSO, IT Dept plans, integration (or not) with user login features, etc. 

I think we should revisit this before APIv3 need authentication features, but that time is not now.